fixed errors and warnings by checking with v-if if the post is already loaded; deleted detailedPost; changed backendurls for server
