+++ frontend/src/router/index.js
@@ -2,7 +2,7 @@ import { createRouter, createWebHistory } from 'vue-router'
 import testEvents from "../views/testEvents.vue"
 import sendAboutUs from "../views/sendAboutUs.vue";
 import News from "../views/News.vue";
-import detailedPost from "../components/detailedPost.vue";
+import postView from "../views/PostView.vue";
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -25,7 +25,7 @@ const router = createRouter({
     {
       path: "/postView:postid(.*)",
       name: "postView",
-      component: detailedPost
+      component: postView
     },
     {
       path: '/about',
